perf(ci): balance process-heavy tooling checks
